% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DE_object.R
\name{Create_DE_object}
\alias{Create_DE_object}
\title{Create_DE_object}
\usage{
Create_DE_object(expr_mtx, sample_info = NULL, gene_info = NULL)
}
\arguments{
\item{expr_mtx}{expression matrix (n genes * p cells) for expression data}

\item{sample_info}{dataframe of p rows, storing sample information.
If it is not specified, a data.frame is created with one
column of colnames of expr_mtx as sample id.}

\item{gene_info}{dataframe of n rows, storing gene annotation information.
If it is not specified, a data.frame is created with one
column of rownames of expr_mtx as gene id.}
}
\value{
a S4 class containing all data and information
}
\description{
Create a object in S4 class for differnetial expression analysis with
  expression matrix, sample information and gene annotation.
}
