% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_test.R
\name{DEA_test}
\alias{DEA_test}
\title{DE_test}
\usage{
DEA_test(x, group, option = c("Wilcox", "T", "NB"), verbose = TRUE)
}
\arguments{
\item{x}{a DE object from Normalize_data}

\item{group}{a colname from sample_info or a vetor of factor indexing two group for comparison}

\item{option}{statistical method to perform differential expression test}

\item{verbose}{Print the information of reference and process}
}
\value{
add a DE_info slot with gene information, log fold change and p value
}
\description{
Perform differential expression analysis
}
