% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_DEG.R
\name{Get_DEG}
\alias{Get_DEG}
\title{Get_DEG}
\usage{
Get_DEG(x, criteria = c("p", "logfc"), threshold, is.sort = T)
}
\arguments{
\item{x}{a DE object containing results from DEA_test}

\item{criteria}{criteria to filter DEG, with options of p value or absolute value of log_FC}

\item{threshold}{threshold of the criteria to filter DEG}

\item{is.sort}{whether sorting genes based on criteria}
}
\value{
a DE object with DE_info containing interesting gene
}
\description{
Get differentail expression genes based on optional criteria
}
