% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{T_Test}
\alias{T_Test}
\title{Differentail Expression Analysis using Student T-test}
\usage{
T_Test = function(data1, data2)
}
\arguments{
\item{data1}{the reference matrix (m genes * n1 samples) for Wilcoxon Rank Sum Test}

\item{data2}{the testing matrix (m genes * n2 samples) for Wilcoxon Rank SUm Test}
}
\description{
Given two group of expression data, returns a vector of p values for each gene (row)
}
\details{

}
