% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{WilcoxTest}
\alias{WilcoxTest}
\title{Differnetial Expression Analysis using Wilcoxon Rank Sum Test}
\usage{
WilcoxTest = function(data1, data2)
}
\arguments{
\item{data1}{the reference matrix (m genes * n1 samples) for Wilcoxon Rank Sum Test}

\item{data2}{the testing matrix (m genes * n2 samples) for Wilcoxon Rank SUm Test}
}
\description{
Given two group of expression data, returns a vector of p values for each gene (row)
}
\details{

}
