% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{WilcoxTest_Vec}
\alias{WilcoxTest_Vec}
\title{Wilcox rank sum test on two vector}
\usage{
WilcoxTest_Vec(x, y)
}
\arguments{
\item{x}{reference vector for Wilcox rank sum test}

\item{y}{testing vector for Wilcox rank sum test}
}
\description{
Given two vector, returns p value for Wilcox rank sum test
}
